LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/multiple-aligned-targets/prefer-targeted-element-expected.txt


FAIL scroller selects targeted area box1 among multiple aligned areas. assert_equals: scroller followed box1 in y axis after layout change expected 100 but got 0
FAIL scroller selects targeted area box2 among multiple aligned areas. assert_equals: scroller followed box2 in y axis after layout change expected 100 but got 0
FAIL scroller selects targeted area box3 among multiple aligned areas. assert_equals: scroller followed box3 in y axis after layout change expected 100 but got 0
FAIL scroller selects targeted area box4 among multiple aligned areas. assert_equals: scroller followed box4 in y axis after layout change expected 300 but got 100
FAIL scroller selects targeted area box5 among multiple aligned areas. assert_equals: scroller followed box5 in y axis after layout change expected 300 but got 100
FAIL scroller selects targeted area box6 among multiple aligned areas. assert_equals: scroller followed box6 in y axis after layout change expected 300 but got 100
FAIL scroller selects targeted area box7 among multiple aligned areas. assert_equals: scroller followed box7 in y axis after layout change expected 500 but got 200
FAIL scroller selects targeted area box8 among multiple aligned areas. assert_equals: scroller followed box8 in y axis after layout change expected 500 but got 200
FAIL scroller selects targeted area box9 among multiple aligned areas. assert_equals: scroller followed box9 in y axis after layout change expected 500 but got 200
